relocate socket-io related modules
