+++ apps/app/src/server/service/socket-io/index.ts
@@ -1,0 +1,1 @@
+export * from './socket-io';
